extract metadata editor and add to file menu
